backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
pytesseract==0.3.10
easyocr==1.7.0
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.200
rembg==2.0.50
scikit-learn==1.3.2
scikit-image==0.22.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1
aiofiles==23.2.1
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4